src/third_party/blink/web_tests/external/wpt/mixed-content/gen/sharedworker-module-data.meta/unset/fetch/cross-http.swap-scheme.https-expected.txt

This is a testharness.js-based test.
FAIL Mixed-Content: Expects blocked for fetch to cross-http origin and swap-scheme redirection from https context. assert_equals: The resource request should be 'blocked'. expected "blocked" but got "allowed"
Harness: the test ran to completion.
